Plug doltcli in
