+++ doltpy/cli/__init__.py
@@ -1,1 +1,1 @@
-from .dolt import Dolt, DoltCommit, DoltException, DoltHubContext
+from doltcli import Dolt, Commit, DoltException, DoltHubContext  # type: ignore
