--- doltpy/cli/read/__init__.py
@@ -1,8 +1,0 @@
-from .read import (
-    read_columns,
-    read_columns_sql,
-    read_pandas,
-    read_pandas_sql,
-    read_rows,
-    read_rows_sql,
-)
