+++ pyproject.toml
@@ -24,6 +24,7 @@ SQLAlchemy = "1.3.18"
 wcwidth = "0.2.5"
 more-itertools = "^8.6.0"
 mysql-connector-python = "^8.0.20"
+doltcli = "^0.1.0"
 
 [tool.poetry.dev-dependencies]
 black = "^20.8b1"
@@ -40,6 +41,7 @@ sphinx_issues = "^1.2.0"
 pytest = "^6.2.2"
 pytest-docker = "^0.10.1"
 pytest-cov = "^2.11.1"
+hypothesis = "^6.7.0"
 
 [tool.poetry.extras]
 pg = ["psycopg2-binary"]
